rename file to cluster.py
